ai-services/requirement.txt
transformers
torch
torchvision
keybert
sentence-transformers
pillow
requests
